# Add live performance counters