… debug/index.html
… src/util/mapbox.js
… src/util/performance.js
… test/unit/ui/map.test.js
… test/unit/util/live_performance.test.js
… test/unit/util/mapbox.test.js