@rafraser 2022-11-14
I'm a little concerned about the lack of communication from Mapbox about this feature. It looks like this has some serious implications for our product's privacy policy.

I understand that this is incredibly valuable data to help improve the product - I'm just not happy that this has been essentially snuck into a release, defaulting to collecting quite detailed data.