@mathewantony31 2022-11-22
@rafraser thanks for flagging your concern, we're updating our release notes to provide additional clarity. Mapbox takes data privacy and security very seriously. In line with our privacy by design practices, the GL JS live performance counter was carefully designed so that user-level metrics and identifiers are not collected. We welcome any follow-on questions you may have regarding our ongoing commitment to data privacy and security.